journal_summarizer/requirements.txt
python-dateutil>=2.8.2
PyYAML>=6.0.1
rich>=13.7.0
